Update  Book Programme 7.py

+++ Book Programme/ Book Programme 7.py
@@ -2,18 +2,19 @@
 #Date = 1st of Semptember 2025
 #description = Program to request user name as input and return the same with a greeting 
 subject_1 =input('Enter Your 1 subject : ')
-subject_1 =int(subject_1)
+subject_1 =float(subject_1)
 subject_2 =input('Enter Your 2 subject : ')
-subject_2 =int(subject_2)
+subject_2 =float(subject_2)
 subject_3 =input('Enter Your 3 subject : ')
-subject_3 = int(subject_3)
+subject_3 = float(subject_3)
 subject_4 =input('Enter Your 4 subject : ')
-subject_4 = int (subject_4)
+subject_4 = float (subject_4)
 subject_5 =input('Enter Your 5 subject : ')
-subject_5 = int(subject_5)
+subject_5 = float(subject_5)
 
 average = (subject_1 + subject_2 + subject_3 + subject_4 + subject_5)/5
 
 print(average)
+
